Move frame tensor allocation APIs to Frame.h
